+++ routes/game_session.js
@@ -67,14 +67,19 @@ module.exports = {
 	postDelete: (req, res) => {
 		let id = req.params.id;
 
-		// db.query to delete game
-		let query = `DELETE FROM GameSessions WHERE game_session_id = ${id}`;
-		db.query(query, (err, result) => {
-			if (err) {
-				throw err;
-			}
-			console.log(`Game Session ${id} deleted`);
+		// db.query to delete game if Yes is pressed
+		if(req.body.delete === 'Yes'){
+			let query = `DELETE FROM GameSessions WHERE game_session_id = ${id}`;
+			db.query(query, (err, result) => {
+				if (err) {
+					throw err;
+				}
+				console.log(`Game Session ${id} deleted`);
+				res.redirect('/');
+			});
+		}
+		else{
 			res.redirect('/');
-		});
+		}
 	}
 };
